refactor the representation evaluation module into submodules
