+++ viscy/representation/evalutation/clustering.py
@@ -1,0 +1,77 @@
+"""Methods for evaluating clustering performance."""
+
+from sklearn.cluster import DBSCAN
+from sklearn.metrics import (
+    accuracy_score,
+    adjusted_rand_score,
+    normalized_mutual_info_score,
+)
+from sklearn.neighbors import KNeighborsClassifier
+
+
+def knn_accuracy(embeddings, annotations, k=5):
+    """
+    Evaluate the k-NN classification accuracy.
+
+    Parameters
+    ----------
+    k : int, optional
+        Number of neighbors to use for k-NN. Default is 5.
+
+    Returns
+    -------
+    float
+        Accuracy of the k-NN classifier.
+    """
+    knn = KNeighborsClassifier(n_neighbors=k)
+    knn.fit(embeddings, annotations)
+    predictions = knn.predict(embeddings)
+    accuracy = accuracy_score(annotations, predictions)
+    return accuracy
+
+
+def dbscan_clustering(embeddings, eps=0.5, min_samples=5):
+    """
+    Apply DBSCAN clustering to the embeddings.
+
+    Parameters
+    ----------
+    eps : float, optional
+        The maximum distance between two samples for them to be considered as in the same neighborhood. Default is 0.5.
+    min_samples : int, optional
+        The number of samples in a neighborhood for a point to be considered as a core point. Default is 5.
+
+    Returns
+    -------
+    np.ndarray
+        Clustering labels assigned by DBSCAN.
+    """
+    dbscan = DBSCAN(eps=eps, min_samples=min_samples)
+    clusters = dbscan.fit_predict(embeddings)
+    return clusters
+
+
+def clustering_evaluation(embeddings, annotations, method="nmi"):
+    """
+    Evaluate the clustering of the embeddings compared to the ground truth labels.
+
+    Parameters
+    ----------
+    method : str, optional
+        Metric to use for evaluation ('nmi' or 'ari'). Default is 'nmi'.
+
+    Returns
+    -------
+    float
+        NMI or ARI score depending on the method chosen.
+    """
+    clusters = dbscan_clustering(embeddings)
+
+    if method == "nmi":
+        score = normalized_mutual_info_score(annotations, clusters)
+    elif method == "ari":
+        score = adjusted_rand_score(annotations, clusters)
+    else:
+        raise ValueError("Invalid method. Choose 'nmi' or 'ari'.")
+
+    return score
